feat: Add .metadata file to track Flutter project properties
This commit adds the `.metadata` file to the project. This file is
used by the Flutter tool to track project properties, assess
capabilities, and perform upgrades.

The file includes information about the Flutter version (revision and
channel), the project type (app), and metadata for the `flutter
migrate` command, specifying platform revisions. It also defines
`unmanaged_files` for the migrate tool (lib/main.dart and
ios/Runner.xcodeproj/project.pbxproj).

This file should be version controlled and not manually edited.