feat: Add IDEWorkspaceChecks.plist to suppress Mac 32-bit warning
This commit adds the `IDEWorkspaceChecks.plist` file to the project.
This file contains the `IDEDidComputeMac32BitWarning` key set to `true`,
which suppresses the "Building a 32-bit application is no longer supported" warning in Xcode when building for macOS.
This warning is irrelevant for Flutter iOS projects and can be safely suppressed to reduce noise in the build process.