feat: Add IDEWorkspaceChecks.plist to suppress 32-bit warning
This commit adds the `IDEWorkspaceChecks.plist` file to the project. This file is used by Xcode to store workspace-specific settings and suppress warnings. In this case, it's being used to suppress the `IDEDidComputeMac32BitWarning`, preventing Xcode from displaying a warning related to 32-bit architecture support on macOS. This warning is not relevant for modern Flutter projects.