feat: Add LaunchImage.png to assets for initial splash screen
This commit adds the `LaunchImage.png` file to the `Assets.xcassets/LaunchImage.imageset` directory within the iOS Runner project. This image will be used as the application's launch image (splash screen) when the app is first opened. This provides a visual indication to the user that the application is loading. The addition of this file is part of the initial project setup to improve user experience.