feat: Add initial Main.storyboard for Flutter integration
This commit introduces the `Main.storyboard` file, which is
necessary for integrating the Flutter application into the iOS
environment. The storyboard defines a single `FlutterViewController`
as the initial view controller.  The background color is set to white.
This setup provides the basic structure for the Flutter UI to be
displayed within the native iOS application.