feat: Initial plugin registration support for Linux
This commit introduces the `generated_plugin_registrant.cc` file,
which provides the basic structure for registering Flutter plugins
on Linux. Currently, the `fl_register_plugins` function is empty,
meaning no plugins are registered by default. This file serves as
a placeholder and will be populated with plugin registration code
generated by Flutter tooling as plugins are added to the project.