feat: Register url_launcher_linux plugin for Linux platform
This commit adds the registration of the `url_launcher_linux` plugin
to the `fl_register_plugins` function in
`generated_plugin_registrant.cc`. This ensures that the url_launcher
plugin is properly initialized and available for use on the Linux
platform within Flutter applications.  The registrar is retrieved and
then used to register the plugin using
`url_launcher_plugin_register_with_registrar`.