feat(linux): Add url_launcher_linux as a generated plugin
Adds `url_launcher_linux` to `FLUTTER_PLUGIN_LIST` in
`generated_plugins.cmake`. This ensures that the `url_launcher_linux`
plugin is properly included and linked during the build process for
Flutter Linux applications. This change enables the functionality
provided by the `url_launcher_linux` plugin to be available in the
application.