feat: Suppress Xcode 32-bit architecture warning
This commit adds `IDEWorkspaceChecks.plist` to suppress the Xcode warning related to 32-bit architecture support on macOS. This warning is irrelevant for Flutter projects, as Flutter does not target 32-bit macOS. Adding this file prevents the warning from appearing during project builds and analysis, improving developer experience. The `IDEDidComputeMac32BitWarning` key is set to `true` to achieve this suppression.