feat: Suppress Xcode 32-bit architecture warning
This commit adds the `IDEWorkspaceChecks.plist` file to the project,
specifically setting the `IDEDidComputeMac32BitWarning` key to `true`.
This suppresses the Xcode warning about the project not supporting 32-bit
architectures on macOS.  This warning is unnecessary as Flutter no
longer supports 32-bit macOS.  This avoids unnecessary UI clutter
in Xcode.