feat: Add maskable 512x512 icon for PWA support
This commit adds the `Icon-maskable-512.png` file to the `web/icons` directory. This is a maskable icon, designed to be adaptable for Progressive Web Apps (PWAs) on various platforms. The 512x512 size ensures compatibility with most app stores and platforms.  This allows the icon to be used flexibly, fitting different shapes and sizes dictated by the operating system.