feat: Add application icon for Windows runner
This commit adds the `app_icon.ico` file to the `windows/runner/resources/` directory. This icon will be used as the application icon for the Flutter application when running on Windows.  Including a dedicated icon improves the user experience by providing a visual identifier for the application in the taskbar and window manager.